Location sharing - extract isSelfLocation util (#8279)
* extract isSelfLocation into utils


* replace use of isSelfLocation
